work with navigation
