+++ src/index.tsx
@@ -3,12 +3,14 @@ import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
 
+import { NavigationProvider } from "./context/navigation";
+
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
 );
 
 root.render(
-  <React.StrictMode>
+  <NavigationProvider>
     <App />
-  </React.StrictMode>
+  </NavigationProvider>
 );
